rearranging code to match better hierarchy
