+++ Extensions/Announcers/GithubAnnouncements/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Newtonsoft.Json" version="4.0.5" />
+</packages>
