+++ Extensions/Sprockets/IPityTheFoolSprocket/PityTheFoolSprocket.cs
