+++ Extensions/Sprockets/IPityTheFoolSprocket/Properties/AssemblyInfo.cs
