+++ Extensions/Sprockets/SampleSprocket/Properties/AssemblyInfo.cs
